+++ expected/wasm32-wasi-threads/undefined-symbols.txt
@@ -62,6 +62,7 @@ __imported_wasi_snapshot_preview1_sock_shutdown
 __imported_wasi_thread_spawn
 __letf2
 __lttf2
+__main_argc_argv
 __netf2
 __stack_pointer
 __subtf3
